TestLevel_1
We don't have to use this at all, but I got  a great script for the camera motion. The idea behid this level was that the player would need to grab the sun from the sky (turing the enviroment to night) to unlock a door. I have yet to complete it, but I'd like some feedback first maybe.