Improved Test Level
I just want to try to successfully commit. Anyway, not sure if this was correctly uploaded. I made a scene in the assets with my name in it. The basis of it was a forrest where the player would take the sun out of the sky(turning day into night) and using it to solve a puzzle. It's not done yet.. We don't have to use it at all if it doesn't fit in.